requirements.txt
Pillow==10.2.0
PyQRCode==1.2.1
